Add files via upload
This Profiler app created using Java is an application that allows you to create files that you can use to keep track of people and their personal details. It features a GUI using SWING and serialization to save and update each file. Is it polished? No. Could it be refined? Yes! More projects are coming as time permits!